BE: Make DateMapper as a Bean
