+++ pilot-project-back/pom.xml
@@ -137,17 +137,13 @@
                 <configuration>
                     <source>${java.version}</source>
                     <target>${java.version}</target>
+                    <useIncrementalCompilation>false</useIncrementalCompilation>
                     <annotationProcessorPaths>
                         <path>
                             <groupId>org.projectlombok</groupId>
                             <artifactId>lombok</artifactId>
                             <version>${lombok.version}</version>
                         </path>
-                        <!--                        <path>-->
-                        <!--                            <groupId>org.projectlombok</groupId>-->
-                        <!--                            <artifactId>lombok-mapstruct-binding</artifactId>-->
-                        <!--                            <version>0.2.0</version>-->
-                        <!--                        </path>-->
                         <path>
                             <groupId>org.mapstruct</groupId>
                             <artifactId>mapstruct-processor</artifactId>
